Introduce InternalDomainCtTypeFactory
